Moved sendNotification method out to auth.setup module
- Had to inject the mailAPI as a parameter to the method to avoid coupling the service with specific API
